Move convert_to_json to file_functions.py, add test for function.
